adapt trick representation
now, cards from middle left and right player are horizontal
