+++ app/src/main/java/com/github/lucaengel/jass_entials/game/JassRoundActivity.kt
@@ -5,20 +5,12 @@ import android.os.Bundle
 import android.widget.Toast
 import androidx.activity.ComponentActivity
 import androidx.activity.compose.setContent
-import androidx.compose.foundation.Image
-import androidx.compose.foundation.clickable
-import androidx.compose.foundation.layout.Box
-import androidx.compose.foundation.layout.BoxWithConstraints
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.Row
 import androidx.compose.foundation.layout.Spacer
 import androidx.compose.foundation.layout.fillMaxHeight
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.fillMaxWidth
-import androidx.compose.foundation.layout.height
-import androidx.compose.foundation.layout.requiredHeight
-import androidx.compose.foundation.layout.requiredWidth
-import androidx.compose.foundation.layout.width
 import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.Surface
 import androidx.compose.material3.Text
@@ -30,17 +22,14 @@ import androidx.compose.runtime.remember
 import androidx.compose.runtime.setValue
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.platform.LocalConfiguration
 import androidx.compose.ui.platform.LocalContext
-import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.tooling.preview.Devices
 import androidx.compose.ui.tooling.preview.Preview
-import androidx.compose.ui.unit.dp
-import androidx.compose.ui.zIndex
 import com.github.lucaengel.jass_entials.data.cards.Card
 import com.github.lucaengel.jass_entials.data.cards.PlayerData
 import com.github.lucaengel.jass_entials.data.game_state.GameState
 import com.github.lucaengel.jass_entials.data.game_state.GameStateHolder
+import com.github.lucaengel.jass_entials.game.JassComposables.Companion.CurrentTrick
 import com.github.lucaengel.jass_entials.game.player.CpuPlayer
 import com.github.lucaengel.jass_entials.game.postgame.SidiBarahniPostRoundActivity
 import com.github.lucaengel.jass_entials.ui.theme.JassentialsTheme
@@ -258,93 +247,6 @@ private fun JassMiddleRowInfo(gameState: GameState, players: List<PlayerData>, c
                 .fillMaxWidth(0.25F)
                 .weight(1f)
         )
-    }
-}
-
-/**
- * Displays the 0-4 cards that are currently in the middle of the table.
- *
- * @param gameState The current game state.
- **/
-@Composable
-fun CurrentTrick(gameState: GameState, players: List<PlayerData>, currentUserIdx: Int, onClick: () -> Unit) {
-
-    val currentTrick = gameState.currentTrick
-    val startingPlayerIdx = gameState.playerEmails.indexOfFirst { it == gameState.startingPlayerEmail }
-
-    // 0 is bottom, 1 is right, 2 is top, 3 is left
-    val idxToCards = (0..3)
-            // Triple: (location of card, i.e. 0 is bottom, ..., index of player in [players], z-index)
-        .map { idx -> Triple(idx, (currentUserIdx + idx) % 4, Math.floorMod(idx - startingPlayerIdx, 4)) }
-        .associate { (i, playerIdx, zIndex) -> i to Pair(currentTrick.trickCards.firstOrNull { it.email == players[playerIdx].email }?.card, zIndex) }
-
-    val screenWidth = LocalConfiguration.current.screenWidthDp.dp
-    val cardWidth = screenWidth.value / 10
-    val cardHeight = cardWidth * 1.5f
-
-    // Displays the cards in the middle of the table.
-    BoxWithConstraints(
-        modifier = Modifier.clickable(onClick = onClick),
-        contentAlignment = Alignment.TopCenter
-    ) {
-        // Card at the top middle
-        Column(modifier = Modifier.zIndex(idxToCards[2]?.second?.toFloat() ?: 0f)) {
-            CardBox(card = idxToCards[2]?.first)
-        }
-
-        // Card at the bottom middle
-        Column(modifier = Modifier.zIndex(idxToCards[0]?.second?.toFloat() ?: 0f)) {
-            Spacer(modifier = Modifier.height((cardHeight * 2 / 3).dp))
-
-            CardBox(card = idxToCards[0]?.first)
-        }
-
-        // Card in the middle on the left
-        Column(modifier = Modifier.zIndex(idxToCards[3]?.second?.toFloat() ?: 0f)) {
-            Spacer(modifier = Modifier.height((cardHeight / 3).dp))
-
-            Row {
-                CardBox(card = idxToCards[3]?.first)
-
-                Spacer(modifier = Modifier.width((cardWidth * 1.25f).dp))
-            }
-        }
-
-        // Card in the middle on the right
-        Column(modifier = Modifier.zIndex(idxToCards[1]?.second?.toFloat() ?: 0f)){
-            Spacer(modifier = Modifier.height((cardHeight / 3).dp))
-
-            Row() {
-                Spacer(modifier = Modifier.width((cardWidth * 1.25f).dp))
-
-                CardBox(card = idxToCards[1]?.first)
-            }
-        }
-    }
-}
-
-/**
- * Displays a jass card image for the given card.
- */
-@Composable
-fun CardBox(card: Card?) {
-    val screenWidth = LocalConfiguration.current.screenWidthDp.dp
-    val cardWidth = screenWidth.value / 10 //* 1.5f
-    val cardHeight = cardWidth * 1.5f
-
-    Box(modifier = Modifier
-        .requiredWidth(cardWidth.dp)
-        .requiredHeight(cardHeight.dp)
-    ) {
-        if (card != null) {
-            Image(
-                painter = painterResource(id = Card.getCardImage(card)),
-                contentDescription = card.toString(),
-                modifier = Modifier
-                    .fillMaxHeight()
-                    .fillMaxSize(),
-            )
-        }
     }
 }
